+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.1.0
 
 *   Implicit / automatic conversions between units of different ratio can be disabled with the preprocessor flag `DISABLE_IMPLICIT_RATIO_CONVERSION`
+*   Fixed division of units with same exponent, so it considers ratio
+*   Bugfix in division of units with different exponent and different ratios
 
 ## 1.0.1